FUNC: Implement better file pattern matcher
